Move views/channel.{py,kv} -> views/conversation_list.{py,kv}
