+++ mephenger/views/conversation_list.py
